# src/main/java/com/activity/activitycalendar/model/UserDto.java
package com.activity.activitycalendar.model;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
@JsonIgnoreProperties(ignoreUnknown = true)
public class UserDto {

    private Long userId;
    private String firstName;
    private String lastName;
    private String userName;
    private String email;
    private String password;
}